microcluster/types: Expose rest related types to allow direct exposure
