+++ microcluster/types/hooks.go
@@ -1,7 +1,5 @@
 package types
 
-import "github.com/canonical/microcluster/v3/rest/types"
-
 // HookType represents the various types of hooks available to microcluster.
 type HookType string
 
@@ -48,5 +46,5 @@ type HookRemoveMemberOptions struct {
 // HookNewMemberOptions holds configuration pertaining to the OnNewMember hook.
 type HookNewMemberOptions struct {
 	// Name is the name of the new cluster member that joined the cluster, triggering this hook.
-	NewMember types.ClusterMemberLocal `json:"new_member" yaml:"new_member"`
+	NewMember ClusterMemberLocal `json:"new_member" yaml:"new_member"`
 }
